global package support. update docs. minor updates and fixes

+++ src/scripts/update_deps.js
@@ -1,16 +1,30 @@
-const { PATHS } = require('../constants')
-const { dependencies, devDependencies } = require(PATHS.package)
+const { PATHS }         = require('../constants')
+const { existsSync }    = require('fs')
 
 
-const depsKeys = dependencies && Object.keys(dependencies).join(' ')
-const devDepsKeys = devDependencies && Object.keys(devDependencies).join(' ')
-
-if (depsKeys || devDepsKeys) {
-    const shell = require('child_process').execSync;
-    const shellOpts = { stdio: 'inherit' }
-
-    shell(`npm remove ${depsKeys} ${devDepsKeys}`, shellOpts)
-
-    depsKeys && shell(`npm i ${depsKeys}`, shellOpts)
-    devDepsKeys && shell(`npm i -D ${devDepsKeys}`, shellOpts)
+function main() {
+    if (existsSync(PATHS.cwdPackageJSON)) {
+        const { dependencies, devDependencies } = require(PATHS.cwdPackageJSON)
+        
+        const depsKeys = dependencies && Object.keys(dependencies).join(' ')
+        const devDepsKeys = devDependencies && Object.keys(devDependencies).join(' ')
+        
+        if (depsKeys || devDepsKeys) {
+            const shell = require('child_process').execSync;
+            const shellOpts = { stdio: 'inherit' }
+        
+            shell(`npm remove ${depsKeys} ${devDepsKeys}`, shellOpts)
+        
+            depsKeys && shell(`npm i ${depsKeys}`, shellOpts)
+            devDepsKeys && shell(`npm i -D ${devDepsKeys}`, shellOpts)
+        }
+    } else {
+        console.error('%s does not exists.', PATHS.cwdPackageJSON)
+        process.exitCode = 1
+    }
 }
+
+
+module.parent
+    ?   (module.exports = main)
+    :   main()
